requirements.txt
Pillow~=10.0.0
Requests~=2.31.0
urllib3~=1.26.15
